refactor(error): finish extraction of ErrorCategory/response/context from god object
exception_handler.py defined ErrorCategory, ErrorCode,
InfrastructureErrorResponse and ExceptionContext inline, duplicating the
categories/responses/context modules scaffolded at MVP but never wired.
Reconcile the DUPLICATE category + its CONFLICT status mapping into the
standalone modules, repoint exception_handler and its consumers to import
from them, and surface error/utilities on the package. Drops
exception_handler.py from ~942 to ~580 LOC with no behavior change.
